FastApiAuth/src/requirements.txt
aioredis==2.0.1
uvicorn==0.19.0
SQLAlchemy==1.4.45
psycopg2==2.9.5
asyncpg==0.27.0
alembic==1.9.0
fastapi==0.89.1
flake8==6.0.0
pydantic==1.10.4
pytest==7.2.1
pytest-env==0.8.1
pbkdf2==1.3
PyJWT==2.6.0
python-dotenv==1.0.0
gunicorn==20.1.0